+++ controllers/ctrMovie.js
@@ -1,5 +1,6 @@
 const { ObjectID } = require('mongodb');
 const { errorHandler } = require('../utility/errorHandler');
+const ctrAuth = require('./ctrAuth');
 const Db = require('../utility/db');
 const config = require('../config');
 const Helper = require('../utility/helper');
@@ -8,19 +9,21 @@ const collectionName = 'movie';
 
 class CtrMovie {
 
-    static async getAll() {
+    static async getAll(global, { token }) {
         let resError = {
             ...config.messages.getFail,
-            data: [
-                {
-                    _id: null,
-                    idDistributor: null,
-                    idsMovieFormat: null,
-                    nombre: null,
-                    active: null
-                }
-            ]
+            data: null
         };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token, option: collectionName, action: config.actions.get });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+
         try {
             let objResult = await Db.find({
                 dbName: config.db.programacion,
@@ -48,15 +51,30 @@ class CtrMovie {
         }
     }
 
-    static async add(root, args) {
-        let exist = await Helper.validateIfExist(
-            {
-                dbName: config.db.programacion,
-                collectionName,
-                params: {
-                    nombre: args.input.nombre
-                }
-            });
+    static async add(global, args) {
+
+        // Build object error.
+        let resError = {
+            ...config.messages.addFail,
+            _id: null
+        };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.add });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+
+        let exist = await Helper.validateIfExist({
+            dbName: config.db.programacion,
+            collectionName,
+            params: {
+                nombre: args.input.nombre
+            }
+        });
         if (exist) {
             return {
                 status: false,
@@ -64,10 +82,6 @@ class CtrMovie {
                 _id: null
             };
         }
-        let resError = {
-            ...config.messages.addFail,
-            _id: null
-        };
         try {
             let newObj = {
                 ...args.input,
@@ -104,17 +118,25 @@ class CtrMovie {
         }
     }
 
-    static async update(root, input) {
+    static async update(global, args) {
         try {
-            if (input.input.idDistributor)
-                input.input.idDistributor = ObjectID(input.input.idDistributor);
-            if (input.input.idsMovieFormat)
-                input.input.idsMovieFormat = input.input.idsMovieFormat.map(m => ObjectID(m));
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.update });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
+            if (args.input.idDistributor)
+                args.input.idDistributor = ObjectID(args.input.idDistributor);
+            if (args.input.idsMovieFormat)
+                args.input.idsMovieFormat = args.input.idsMovieFormat.map(m => ObjectID(m));
             let objResult = await Db.update({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id,
-                set: { ...input.input, updated_at: new Date() }
+                _id: args._id,
+                set: { ...args.input, updated_at: new Date() }
             });
             if (!objResult.status) {
                 errorHandler({
@@ -136,12 +158,20 @@ class CtrMovie {
         }
     }
 
-    static async delete(root, input) {
+    static async delete(global, args) {
         try {
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.delete });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
             let objResult = await Db.delete({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id
+                _id: args._id
             });
             if (!objResult.status) {
                 errorHandler({
