+++ controllers/ctrRoom.js
@@ -1,5 +1,6 @@
 const { ObjectID } = require('mongodb');
 const { errorHandler } = require('../utility/errorHandler');
+const ctrAuth = require('./ctrAuth');
 const Db = require('../utility/db');
 const config = require('../config');
 const Helper = require('../utility/helper');
@@ -8,20 +9,23 @@ const collectionName = 'room';
 
 class CtrRoom {
 
-    static async getAll() {
+    static async getAll(global, { token }) {
+
+        // Build object error.
         let resError = {
             ...config.messages.getFail,
-            data: [
-                {
-                    _id: null,
-                    idTheater: null,
-                    nombre: null,
-                    numberChairs: null,
-                    cleaningTime: null,
-                    active: null
-                }
-            ]
+            data: null
         };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token, option: collectionName, action: config.actions.get });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+
         try {
             let objResult = await Db.find({
                 dbName: config.db.programacion,
@@ -49,13 +53,28 @@ class CtrRoom {
         }
     }
 
-    static async add(root, args) {
-        let exist = await Helper.validateIfExist(
-            {
-                dbName: config.db.programacion,
-                collectionName,
-                params: { nombre: args.input.nombre, idTheater: ObjectID(args.input.idTheater) }
-            });
+    static async add(global, args) {
+
+        // Build object error.
+        let resError = {
+            ...config.messages.addFail,
+            _id: null
+        };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.add });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+
+        let exist = await Helper.validateIfExist({
+            dbName: config.db.programacion,
+            collectionName,
+            params: { nombre: args.input.nombre, idTheater: ObjectID(args.input.idTheater) }
+        });
         if (exist) {
             return {
                 status: false,
@@ -63,10 +82,6 @@ class CtrRoom {
                 _id: null
             };
         }
-        let resError = {
-            ...config.messages.addFail,
-            _id: null
-        };
         try {
             let newObj = {
                 ...args.input,
@@ -102,15 +117,23 @@ class CtrRoom {
         }
     }
 
-    static async update(root, input) {
+    static async update(global, args) {
         try {
-            if (input.input.idTheater)
-                input.input.idTheater = ObjectID(input.input.idTheater);
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.update });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
+            if (args.input.idTheater)
+                args.input.idTheater = ObjectID(args.input.idTheater);
             let objResult = await Db.update({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id,
-                set: { ...input.input, updated_at: new Date() }
+                _id: args._id,
+                set: { ...args.input, updated_at: new Date() }
             });
             if (!objResult.status) {
                 errorHandler({
@@ -132,12 +155,20 @@ class CtrRoom {
         }
     }
 
-    static async delete(root, input) {
+    static async delete(global, args) {
         try {
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.delete });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
             let objResult = await Db.delete({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id
+                _id: args._id
             });
             if (!objResult.status) {
                 errorHandler({
