+++ controllers/ctrRoomType.js
@@ -1,4 +1,5 @@
 const { errorHandler } = require('../utility/errorHandler');
+const ctrAuth = require('./ctrAuth');
 const Db = require('../utility/db');
 const config = require('../config');
 const Helper = require('../utility/helper');
@@ -7,17 +8,23 @@ const collectionName = 'room_type';
 
 class CtrRoomType {
 
-    static async getAll() {
+    static async getAll(global, { token }) {
+
+        // Build object error.
         let resError = {
             ...config.messages.getFail,
-            data: [
-                {
-                    _id: null,
-                    nombre: null,
-                    active: null
-                }
-            ]
+            data: null
         };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token, option: collectionName, action: config.actions.get });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+
         try {
             let objResult = await Db.find({
                 dbName: config.db.programacion,
@@ -45,13 +52,28 @@ class CtrRoomType {
         }
     }
 
-    static async add(root, args) {
-        let exist = await Helper.validateIfExist(
-            {
-                dbName: config.db.programacion,
-                collectionName,
-                params: { nombre: args.input.nombre }
-            });
+    static async add(global, args) {
+
+        // Build object error.
+        let resError = {
+            ...config.messages.addFail,
+            _id: null
+        };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.add });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+
+        let exist = await Helper.validateIfExist({
+            dbName: config.db.programacion,
+            collectionName,
+            params: { nombre: args.input.nombre }
+        });
         if (exist) {
             return {
                 status: false,
@@ -59,10 +81,6 @@ class CtrRoomType {
                 _id: null
             };
         }
-        let resError = {
-            ...config.messages.addFail,
-            _id: null
-        };
         try {
             let newObj = {
                 ...args.input,
@@ -97,13 +115,21 @@ class CtrRoomType {
         }
     }
 
-    static async update(root, input) {
+    static async update(global, args) {
         try {
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.update });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
             let objResult = await Db.update({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id,
-                set: { ...input.input, updated_at: new Date() }
+                _id: args._id,
+                set: { ...args.input, updated_at: new Date() }
             });
             if (!objResult.status) {
                 errorHandler({
@@ -125,12 +151,20 @@ class CtrRoomType {
         }
     }
 
-    static async delete(root, input) {
+    static async delete(global, args) {
         try {
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.delete });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
             let objResult = await Db.delete({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id
+                _id: args._id
             });
             if (!objResult.status) {
                 errorHandler({
