+++ controllers/ctrUserRol.js
@@ -1,5 +1,6 @@
 const { ObjectID } = require('mongodb');
 const { errorHandler } = require('../utility/errorHandler');
+const ctrAuth = require('./ctrAuth');
 const Db = require('../utility/db');
 const config = require('../config');
 const Helper = require('../utility/helper');
@@ -8,18 +9,21 @@ const collectionName = 'user_rol';
 
 class CtrUserRol {
 
-    static async getAll() {
+    static async getAll(global, { token }) {
         let resError = {
             ...config.messages.getFail,
-            data: [
-                {
-                    _id: null,
-                    nombre: null,
-                    permissions: null,
-                    active: null
-                }
-            ]
+            data: null
         };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token, option: collectionName, action: config.actions.get });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+
         try {
             let objResult = await Db.find({
                 dbName: config.db.programacion,
@@ -47,15 +51,29 @@ class CtrUserRol {
         }
     }
 
-    static async add(root, args) {
-        let exist = await Helper.validateIfExist(
-            {
-                dbName: config.db.programacion,
-                collectionName,
-                params: {
-                    nombre: args.input.nombre
-                }
-            });
+    static async add(global, args) {
+
+        // Build object error.
+        let resError = {
+            ...config.messages.addFail,
+            _id: null
+        };
+
+        // Validation permissions.
+        let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.add });
+        if (!auth.status) {
+            return {
+                ...resError,
+                message: auth.message
+            };
+        }
+        let exist = await Helper.validateIfExist({
+            dbName: config.db.programacion,
+            collectionName,
+            params: {
+                nombre: args.input.nombre
+            }
+        });
         if (exist) {
             return {
                 status: false,
@@ -63,10 +81,6 @@ class CtrUserRol {
                 _id: null
             };
         }
-        let resError = {
-            ...config.messages.addFail,
-            _id: null
-        };
         try {
             let permissions = args.input.permissions.map(m => {
                 return {
@@ -108,10 +122,18 @@ class CtrUserRol {
         }
     }
 
-    static async update(root, input) {
+    static async update(global, args) {
         try {
-            if (input.input.permissions)
-                input.input.permissions = input.input.permissions.map(m => {
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.update });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
+            if (args.input.permissions)
+                args.input.permissions = args.input.permissions.map(m => {
                     return {
                         ...m,
                         idUserOption: ObjectID(m.idUserOption)
@@ -121,8 +143,8 @@ class CtrUserRol {
             let objResult = await Db.update({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id,
-                set: { ...input.input, updated_at: new Date() }
+                _id: args._id,
+                set: { ...args.input, updated_at: new Date() }
             });
             if (!objResult.status) {
                 errorHandler({
@@ -144,12 +166,20 @@ class CtrUserRol {
         }
     }
 
-    static async delete(root, input) {
+    static async delete(global, args) {
         try {
+            // Validation permissions.
+            let auth = ctrAuth.validateLogin({ token: args.token, option: collectionName, action: config.actions.delete });
+            if (!auth.status) {
+                return {
+                    status: false,
+                    message: auth.message
+                };
+            }
             let objResult = await Db.delete({
                 dbName: config.db.programacion,
                 collectionName,
-                _id: input._id
+                _id: args._id
             });
             if (!objResult.status) {
                 errorHandler({
